requirements.txt
Django==3.1.7
sorl_thumbnail==12.7.0
django-debug-toolbar==3.2
Pillow==8.2.0
djangorestframework==3.12.4